error C2593: 'operator <<' is ambiguous http://support.microsoft.com/kb/168440

+++ 3rdparty/rpcz/src/rpcz/zmq_utils.cc
@@ -134,13 +134,13 @@ bool forward_message(zmq::socket_t &socket_in,
 }
 
 void log_message_vector(message_vector& vector) {
-  LOG(INFO) << "---- " << vector.size() << "----";
+  LOG(INFO) << "---- " << static_cast<int>(vector.size()) << "----";
   boost::hash<std::string> hash;
   for (int i = 0; i < vector.size(); ++i) {
     std::string message(message_to_string(vector[i]));
     std::stringstream ss;
     ss << std::hex << hash(message);
-    LOG(INFO) << "(" << vector[i].size() << "): " << "[" << ss.str()
+    LOG(INFO) << "(" << static_cast<int>(vector[i].size()) << "): " << "[" << ss.str()
         << "]: "
         << message_to_string(vector[i]);
   }
